Engine afrondings problemen opgelost!

+++ BeoordelingProject/BeoordelingProject/Engine/IBeoordelingsEngine.cs
@@ -4,6 +4,6 @@ namespace BeoordelingProject.Engine {
         double deelaspect(double midden, int weging);
         double totaalDeelaspect(System.Collections.Generic.List<double> middens, System.Collections.Generic.List<int> wegingen);
         double totaalScore(System.Collections.Generic.List<double> middens, System.Collections.Generic.List<int> wegingen);
-        double totaalWeging(System.Collections.Generic.List<int> wegingen);
+        int totaalWeging(System.Collections.Generic.List<int> wegingen);
     }
 }
